+++ lib/src/widget/impl/facade.dart
@@ -3,7 +3,9 @@ import 'package:cloudflare_turnstile/src/widget/interface.dart' as i;
 import 'package:cloudflare_turnstile/src/widget/turnstile_options.dart';
 import 'package:flutter/material.dart';
 
-class CloudFlareTurnstile extends StatelessWidget implements i.CloudFlareTurnstile {
+/// Facade class
+class CloudFlareTurnstile extends StatelessWidget
+    implements i.CloudFlareTurnstile {
   /// Create a Cloudflare Turnstile Widget
   CloudFlareTurnstile({
     super.key,
@@ -52,7 +54,7 @@ class CloudFlareTurnstile extends StatelessWidget implements i.CloudFlareTurnsti
   final TurnstileController<dynamic>? controller;
 
   /// A Callback invoked upon success of the challange.
-  /// The callback is passed a [token] that can be validated.
+  /// The callback is passed a `token` that can be validated.
   ///
   /// example:
   /// ```dart
@@ -100,6 +102,7 @@ class CloudFlareTurnstile extends StatelessWidget implements i.CloudFlareTurnsti
 
   @override
   Widget build(BuildContext context) {
-    throw UnimplementedError('Cannot call build on the facade implementation of CloudFlareTurnstile.');
+    throw UnimplementedError(
+        'Cannot call build on the facade implementation of CloudFlareTurnstile.');
   }
 }