+++ lib/src/widget/turnstile_options.dart
@@ -10,19 +10,23 @@ class TurnstileOptions {
     this.refreshExpired = TurnstileRefreshExpired.auto,
     this.refreshTimeout = TurnstileRefreshTimeout.auto,
   })  : assert(
-          retryInterval.inMilliseconds > 0 && retryInterval.inMilliseconds <= 900000,
+          retryInterval.inMilliseconds > 0 &&
+              retryInterval.inMilliseconds <= 900000,
           'Duration must be greater than 0 and less than or equal to 900000 milliseconds.',
         ),
         assert(
-          !(mode == TurnstileMode.invisible && refreshExpired == TurnstileRefreshExpired.manual),
+          !(mode == TurnstileMode.invisible &&
+              refreshExpired == TurnstileRefreshExpired.manual),
           '$refreshExpired is impossible in $mode, consider using TurnstileRefreshExpired.auto or TurnstileRefreshExpired.never',
         ),
         assert(
-          !(mode == TurnstileMode.invisible && refreshTimeout != TurnstileRefreshTimeout.auto),
+          !(mode == TurnstileMode.invisible &&
+              refreshTimeout != TurnstileRefreshTimeout.auto),
           '$refreshTimeout has no effect on an $mode widget.',
         ),
         assert(
-          !(mode == TurnstileMode.nonInteractive && refreshTimeout != TurnstileRefreshTimeout.auto),
+          !(mode == TurnstileMode.nonInteractive &&
+              refreshTimeout != TurnstileRefreshTimeout.auto),
           '$refreshTimeout has no effect on an $mode widget.',
         );